separando modules das funcoes JS
