+++ efeitos-no-dom/modules/index.html
@@ -131,7 +131,6 @@
     <p>Todos os direitos reservados.</p>
   </footer>
 
-  <script src="animacoes.js"></script>  
-  <script src="script.js"></script>  
+  <script type="module" src="./js/script.js"></script>  
 </body>
 </html>
