+++ efeitos-no-dom/modules/js/modules/animacao-por-scroll.js
@@ -1,0 +1,20 @@
+export default function initAnimacaoScroll() {
+  const sections = document.querySelectorAll('[data-anime="scroll"]');
+
+  if(sections.length) {
+    const windowMetade = window.innerHeight * 0.5;
+
+    function animaScroll(){
+      sections.forEach((section) => {
+        const sectionTop = section.getBoundingClientRect().top;
+        const isSectionVisible = (sectionTop - windowMetade) < 0;
+        // console.log(sectionTop);
+        if(isSectionVisible) {
+          section.classList.add('ativo');
+        }
+      })
+    }
+    animaScroll()
+    window.addEventListener('scroll', animaScroll);
+  }
+}
