+++ efeitos-no-dom/modules/js/modules/scroll-suave.js
@@ -1,0 +1,17 @@
+export default function initScrollSuave(){
+  const linksInternos = document.querySelectorAll('[data-menu="suave"] a[href^="#"]');
+  // console.log(linksInternos)
+
+  function scrollToSection(event) {
+    event.preventDefault();
+    const href = event.currentTarget.getAttribute('href');
+    const section = document.querySelector(href);  
+    section.scrollIntoView({
+      behavior: 'smooth',
+      block: 'start',
+    });
+  }
+  linksInternos.forEach((link) => {
+    link.addEventListener('click', scrollToSection);
+  });
+}
