+++ efeitos-no-dom/modules/js/script.js
@@ -1,0 +1,8 @@
+import initScrollSuave from './modules/scroll-suave.js';
+import initAnimacaoScroll from './modules/animacao-por-scroll.js';
+import initAccordion from './modules/accordion.js';
+import initTabNav from './modules/tab-nav.js';
+initTabNav();
+initScrollSuave();
+initAnimacaoScroll();
+initAccordion();
