--- efeitos-no-dom/modules/script.js
@@ -1,21 +1,0 @@
-// const div = document.querySelector('[data-cor]');
-
-// div.dataset.height = 1000;
-// delete div.dataset.width;
-// console.log(div.dataset)
-
-// Exercícios:
-
-// Adicione um atributo data-anime="show-down" e
-// data-anime="show-right" a todos as section's
-// com descricão dos animais.
-
-// Utilizando estes atributos, adicione a classe
-// show-down ou show-right a sua respectiva section
-// assim que a mesma aparecer na tela (animacao tab)
-
-// No CSS faça com que show-down anime de cima para baixo
-// e show-right continue com a mesma animação da esquerda
-// para a direita
-
-// Substitua todas as classes js- por data atributes.
